second commit

+++ wetherapp/lib/model/weather_forecast_model.dart
@@ -146,7 +146,7 @@ class Lista {
     gust = json['gust'].toDouble();
     clouds = json['clouds'];
     pop = json['pop'].toDouble();
-    rain = json['rain'];
+    rain = json['rain'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
@@ -215,10 +215,10 @@ class FeelsLike {
   FeelsLike({this.day, this.night, this.eve, this.morn});
 
   FeelsLike.fromJson(Map<String, dynamic> json) {
-    day = json['day'];
-    night = json['night'];
-    eve = json['eve'];
-    morn = json['morn'];
+    day = json['day'].toDouble();
+    night = json['night'].toDouble();
+    eve = json['eve'].toDouble();
+    morn = json['morn'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
